refactor(imports): update config import paths for consistency and clarity; add Twitter and trend detection client entry points
