+++ submissions/myproject/package.json
@@ -2,11 +2,11 @@
     "name": "trendsnipper",
     "version": "0.1.0",
     "description": "An AI agent that detects Twitter micro-trends for Subnet 59",
-    "main": "src/index.js",
+    "main": "src/app.js",
     "type": "module",
     "scripts": {
-        "start": "node src/index.js",
-        "dev": "nodemon src/index.js",
+        "start": "node src/app.js",
+        "dev": "nodemon src/app.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "keywords": [
