+++ submissions/myproject/src/ai/openai-client.js
@@ -1,5 +1,5 @@
 import OpenAI from 'openai';
-import config from '../config.js';
+import config from '../config/config.js';
 import logger from '../utils/logger.js';
 
 class OpenAIClient {
