+++ submissions/myproject/src/mcp/client.js
@@ -1,7 +1,8 @@
+// mcp/client.js - Client MCP
 import { Client } from '@modelcontextprotocol/sdk/client/index.js';
 import { StdioClientTransport } from '@modelcontextprotocol/sdk/client/stdio.js';
 import logger from '../utils/logger.js';
-import config from '../config.js';
+import config from '../config/config.js';
 
 class McpClient {
   constructor() {
