+++ submissions/myproject/src/trend-detection/index.js
@@ -1,0 +1,6 @@
+// trend-detection/index.js - Point d'entrée pour le détecteur de tendances
+import TrendDetector from './detector.js';
+
+// Export de l'instance du détecteur de tendances
+const trendDetector = new TrendDetector();
+export default trendDetector;
