+++ submissions/myproject/src/twitter-client/CookieManager.js
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-import logger from './utils/logger.js';
+import logger from '../utils/logger.js';
 
 class CookieManager {
   constructor(cookiePath = './data/cookies.json') {
