+++ submissions/myproject/src/twitter-client/index.js
@@ -1,0 +1,6 @@
+// twitter-client/index.js - Point d'entrée pour le client Twitter
+import TwitterClient from './client.js';
+
+// Export de l'instance du client Twitter
+const twitterClient = new TwitterClient();
+export default twitterClient;
